requirements.txt
aiohttp==3.5.4
aiomysql==0.0.20
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
crypto==1.4.1
cryptography==2.5
idna==2.8
lxml==4.3.1
multidict==4.5.2
Naked==0.1.31
peewee==3.8.2
peewee-async==0.6.0a0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.2
PyYAML==3.13
redis==3.2.0
requests==2.21.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.7.3
tornado==5.1.1
urllib3==1.24.1
WTForms==2.2.1
WTForms-JSON==0.3.3
wtforms-tornado==0.0.2
yarl==1.3.0
